CachedIEnumerable separated into two files (Static vs Object)
